# apps/api/docs/swagger.yaml
basePath: /
definitions:
  auth.LoginDTO:
    properties:
      email:
        example: lucasmadranges@gmail.com
        type: string
      password:
        example: Azertye789456&!
        minLength: 12
        type: string
    required:
      - email
      - password
    type: object
  mainType.CreateMainTypeDto:
    properties:
      name:
        example: Restaurant
        maxLength: 25
        type: string
    required:
      - name
    type: object
  request.ErrorRequest:
    properties:
      explicit:
        type: string
      message:
        type: string
      success:
        example: false
        type: boolean
    type: object
  request.SuccessCreateRequest:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
  request.SuccessDeleteRequest:
    properties:
      count:
        example: 0
        type: integer
      success:
        example: true
        type: boolean
    type: object
  request.SuccessGetAllRequest:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
  request.SuccessGetOneRequest:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
  subType.CreateSubTypeDto:
    properties:
      main_type_id:
        example: 1
        type: integer
      name:
        example: Buffet à volonté
        maxLength: 25
        type: string
    required:
      - main_type_id
      - name
    type: object
  user.CreateUserDTO:
    properties:
      avatar:
        example: https://example.com/
        type: string
      email:
        example: lucasmadranges@gmail.com
        type: string
      firstname:
        example: Lucas
        type: string
      lastname:
        example: Madranges
        type: string
      password:
        example: Azertye789456&!
        minLength: 12
        type: string
      role:
        allOf:
          - $ref: '#/definitions/user.Role'
        enum:
          - admin
          - user
        example: user
    required:
      - email
      - firstname
      - lastname
      - password
    type: object
  user.Role:
    enum:
      - user
      - admin
      - user
    type: string
    x-enum-varnames:
      - DefaultRole
      - RoleAdmin
      - RoleUser
host: localhost:4000
info:
  contact: {}
  description: API avec Fiber, Ent et Swagger
  title: Go Fiber API
  version: '1.0'
paths:
  /auth/login:
    post:
      consumes:
        - application/json
      description: Authenticate user with email and password
      parameters:
        - description: Login credentials
          in: body
          name: credentials
          required: true
          schema:
            $ref: '#/definitions/auth.LoginDTO'
      produces:
        - application/json
      responses:
        '200':
          description: Retourne un JWT token
          schema:
            $ref: '#/definitions/request.SuccessGetOneRequest'
        '400':
          description: Données invalides
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '401':
          description: Données de connexion invalides
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '404':
          description: Utilisateur non trouvé
          schema:
            $ref: '#/definitions/request.ErrorRequest'
      summary: User login
      tags:
        - Auth
  /mainTypes:
    get:
      description: Retrieve a list of all main types
      produces:
        - application/json
      responses:
        '201':
          description: Récupérer toutes les catégories
          schema:
            $ref: '#/definitions/request.SuccessGetAllRequest'
        '400':
          description: Échec de la récupération des catégories
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '500':
          description: Erreur interne
          schema:
            $ref: '#/definitions/request.ErrorRequest'
      summary: Get all main types
      tags:
        - MainType
    post:
      consumes:
        - application/json
      description: Create a new main type
      parameters:
        - description: mainType payload
          in: body
          name: mainType
          required: true
          schema:
            $ref: '#/definitions/mainType.CreateMainTypeDto'
      produces:
        - application/json
      responses:
        '201':
          description: Catégorie créé avec succés
          schema:
            $ref: '#/definitions/request.SuccessCreateRequest'
        '400':
          description: Données invalides
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '409':
          description: Catégorie déjà existante
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '500':
          description: Erreur interne
          schema:
            $ref: '#/definitions/request.ErrorRequest'
      summary: Create mainType
      tags:
        - MainType
  /subTypes:
    get:
      description: Retrieve a list of all sub types
      produces:
        - application/json
      responses:
        '201':
          description: Récupérer toutes les sous-catégories
          schema:
            $ref: '#/definitions/request.SuccessGetAllRequest'
        '400':
          description: Échec de la récupération des sous-catégories
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '500':
          description: Erreur interne
          schema:
            $ref: '#/definitions/request.ErrorRequest'
      summary: Get all subTypes
      tags:
        - SubType
    post:
      consumes:
        - application/json
      description: Create a new sub type
      parameters:
        - description: subType payload
          in: body
          name: subType
          required: true
          schema:
            $ref: '#/definitions/subType.CreateSubTypeDto'
      produces:
        - application/json
      responses:
        '201':
          description: Sous-catégorie créé avec succés
          schema:
            $ref: '#/definitions/request.SuccessCreateRequest'
        '400':
          description: Données invalides
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '409':
          description: Sous-catégorie déjà existante
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '500':
          description: Erreur interne
          schema:
            $ref: '#/definitions/request.ErrorRequest'
      summary: Create subType
      tags:
        - SubType
  /subTypes/{mainTypeId}:
    get:
      description: Retrieve a list of all sub types for a specific main type ID
      parameters:
        - description: MainType ID
          in: path
          name: mainTypeId
          required: true
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: Récupérer toutes les sous-catégories d'une catégorie
          schema:
            $ref: '#/definitions/request.SuccessGetAllRequest'
        '400':
          description: Échec de la récupération des sous-catégories
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '500':
          description: Erreur interne
          schema:
            $ref: '#/definitions/request.ErrorRequest'
      summary: Get all subTypes by main type ID
      tags:
        - SubType
  /users:
    get:
      description: Retrieve a list of all users
      produces:
        - application/json
      responses:
        '201':
          description: Récupérer tous les utilisateurs
          schema:
            $ref: '#/definitions/request.SuccessGetAllRequest'
        '400':
          description: Échec de la récupération des utilisateurs
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '500':
          description: Erreur interne
          schema:
            $ref: '#/definitions/request.ErrorRequest'
      summary: Get all users
      tags:
        - Users
    post:
      consumes:
        - application/json
      description: Create a new user account
      parameters:
        - description: User payload
          in: body
          name: user
          required: true
          schema:
            $ref: '#/definitions/user.CreateUserDTO'
      produces:
        - application/json
      responses:
        '201':
          description: Utilisateur créé avec succés
          schema:
            $ref: '#/definitions/request.SuccessCreateRequest'
        '400':
          description: Données invalides
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '409':
          description: Utilisateur déjà existant
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '500':
          description: Erreur interne
          schema:
            $ref: '#/definitions/request.ErrorRequest'
      summary: Create user
      tags:
        - Users
  /users/{email}:
    delete:
      description: Delete a user account by email
      parameters:
        - description: User Email
          in: path
          name: email
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: Utilisateur supprimé avec succés
          schema:
            $ref: '#/definitions/request.SuccessDeleteRequest'
        '400':
          description: Email invalide
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '404':
          description: Utilisateur non trouvé
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '500':
          description: Erreur interne
          schema:
            $ref: '#/definitions/request.ErrorRequest'
      summary: Delete user by email
      tags:
        - Users
    get:
      description: Retrieve a user account by email
      parameters:
        - description: User Email
          in: path
          name: email
          required: true
          type: string
      produces:
        - application/json
      responses:
        '201':
          description: Récupérer un utilisateur par email
          schema:
            $ref: '#/definitions/request.SuccessGetAllRequest'
        '400':
          description: Échec de la récupération de l'utilisateur
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '404':
          description: Utilisateur non trouvé
          schema:
            $ref: '#/definitions/request.ErrorRequest'
        '500':
          description: Erreur interne
          schema:
            $ref: '#/definitions/request.ErrorRequest'
      summary: Get user by email
      tags:
        - Users
swagger: '2.0'